[FIX] account.move: _move_autocomplete forgot fiscal position and tax
Per documentation, and also for the sanity of programmers, creating a
new invoice programmatically does not require setting any and all
fields, like fiscal position, or the exact accounts and taxes for
each account.move.line.  Just like in the form view, Odoo can and
should autocomplete all information possible, like accounts and taxes,
with the same result as in the web frontend.

Given a partner_id for account.move and a product_id for a line,
such an autocomplete would be possible - but just creating an
account move per dict did not trigger that.

This commit adds calls to the proper methods in such cases, and also
fixes a missing fiscal-position autocomplete for the account.move.
Also some minor refactor to unify the repeatedly called autocomplete.